+++ controllers/TerminalController.php
@@ -4,8 +4,12 @@ namespace wdmg\terminal\controllers;
 
 use Yii;
 use yii\web\Controller;
+use yii\web\Response;
+use yii\helpers\Url;
+use yii\helpers\Json;
 use yii\filters\VerbFilter;
 use yii\filters\AccessControl;
+
 
 /**
  * TerminalController
@@ -62,6 +66,15 @@ class TerminalController extends Controller
     }
 
     /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        Yii::$app->request->enableCsrfValidation = false;
+        parent::init();
+    }
+
+    /**
      * Main index action.
      * @return mixed
      */
@@ -82,10 +95,55 @@ class TerminalController extends Controller
                 $prompt = Yii::$app->user->identity->username . ':'. Yii::$app->request->serverName .' '.$path.'$ ';
         }
 
+        $rpcRoute = Url::toRoute(['terminal/rpc']);
+
         return $this->render('index', [
             'module' => $this->module,
             'greetings' => $greetings,
+            'rpcRoute' => $rpcRoute,
             'prompt' => $prompt
         ]);
+    }
+
+
+    /**
+     * RPC action
+     * @return array
+     */
+    public function actionRpc()
+    {
+        Yii::$app->response->format = Response::FORMAT_JSON;
+        $options = Json::decode(Yii::$app->request->getRawBody());
+
+
+
+        if (intval($options['jsonrpc']) >= 2 && $options['method'] == "system.describe") {
+            list ($status, $output) = $this->runConsole(Yii::getAlias('@app/yii'), implode(' ', $options['params']));
+            return ['result' => $output];
+        }
+    }
+
+
+    /**
+     * Runs console command
+     *
+     * @param string $cmd
+     * @param string $command
+     * @return null or array [status, output]
+     */
+    private function runConsole($cmd, $command)
+    {
+        if ($cmd) {
+            set_time_limit(0);
+            $cmd = Yii::getAlias($cmd) . ' ' . $command . ' 2>&1';
+            $handler = popen($cmd, 'r');
+            $output = '';
+            while (!feof($handler)) {
+                $output .= fgets($handler);
+            }
+            return [pclose($handler), trim($output)];
+        } else {
+            return null;
+        }
     }
 }