+++ views/terminal/index.php
@@ -19,13 +19,35 @@ $bundle = TerminalAsset::register($this);
 <?php $this->registerJs(<<< JS
 jQuery(function($, undefined) {
     var terminal = $('#terminalArea');
-    terminal.terminal(function(command) {
-        if (command !== '') {
-            var result = window.eval(command);
-            if (result != undefined) {
-                this.echo(String(result));
-            }
+    terminal.terminal(function(command, term) {
+        
+        if (command.indexOf('yii') === 0 || command.indexOf('yii') === 3) {
+            term.pause();
+            $.jrpc('{$rpcRoute}', 'system.describe', [command.replace(/^yii ?/, '')], function(response, status, jqXHR) {
+                term.echo(response.result).resume();
+            }, function() {
+                term.resume();
+            });
+        } else if (command === 'help') {
+            term.echo('Available commands are:');
+            term.echo("\tclear\t- to clear the console");
+            term.echo('\thelp\t- print this help dialog');
+            term.echo('\tyii\t\t- list of yii command`s');
+            term.echo('\tquit\t- to quit from terminal');
+            term.echo('');
+        } else if (command === '') {
+            term.resume();
         }
+        term.echo('');
+        
+        $('html, body').animate({
+            scrollTop: terminal.height()
+        }, 'fast');
+        
+        $('html').on('keydown', function(e) {
+            terminal.click();
+        });
+        
     }, {
         greetings: '$greetings',
         name: 'yii2-terminal',
@@ -33,9 +55,6 @@ jQuery(function($, undefined) {
         prompt: '$prompt'
     });
 });
-
-
-
 JS
 ); ?>